+++ src/main/java/com/example/Tozin_Solutions_back_end/dto/movimentacaoEstoqueDTO/RegistrarMovimentacaoDTO.java
@@ -3,21 +3,21 @@ package com.example.Tozin_Solutions_back_end.dto.movimentacaoEstoqueDTO;
 import com.example.Tozin_Solutions_back_end.model.Peca;
 
 public class RegistrarMovimentacaoDTO {
-    private Integer quantidadeSaida;
-    private Integer quantidadeEntrada;
+    private Double quantidadeSaida;
+    private Double quantidadeEntrada;
     private Peca peca;
 
-    public RegistrarMovimentacaoDTO(Integer quantidadeSaida, Integer quantidadeEntrada, Peca peca) {
+    public RegistrarMovimentacaoDTO(Double quantidadeSaida, Double quantidadeEntrada, Peca peca) {
         this.quantidadeSaida = quantidadeSaida;
         this.quantidadeEntrada = quantidadeEntrada;
         this.peca = peca;
     }
 
-    public Integer getQuantidadeSaida() {
+    public Double getQuantidadeSaida() {
         return quantidadeSaida;
     }
 
-    public Integer getQuantidadeEntrada() {
+    public Double getQuantidadeEntrada() {
         return quantidadeEntrada;
     }
